chore: improve gas limit test
